Move effector API to a separate file
